fix: address CodeRabbit repair feedback
